Update create-a-custom-skin.md
Several grammar changes. I'm unfamiliar with how you are using the word coined in the text, "the wrapping skin-specific class is coined by" and I wonder if you can rewrite that to say created instead. I think that you want to say that one creates the name for the new skin by the control, plus underscore ("_") plus SkinName... Is that correct?